Split 02_get_data into two scripts; download gzips
For some reason, the attempts to download the data from the FTP server
is intermitten on Travis. I initially thought there was some missing
package or setting, perhaps configuration. But when I debugged into a
Travis session I had a few instances where I could use "curl" from the
terminal with success. I had also seen this issue in the "rrricanes"
package.

So, and maybe not a bad alternative, I decided to just download all
the gzips into the data directory. This won't give much speed in the
processing but does open some alternatives (such as if you only want
to examine a specific year for the details dataset. There is no
limit to the size of a repo so this will not require LFS.

TODO
The 02_download_data.R script does not cross check existing files in
the FTP directory with that in the data directory; so it will need
to be modified to only check for updates. And I am unsure if it will
automatically overwrite or throw an error (no parameter and no
mention in the docs).

The script should be modified to only look for updates somehow.